+++ scripts/database/otrs-schema.postgresql.sql
@@ -49,7 +49,7 @@ CREATE TABLE valid (
 CREATE TABLE users (
     id serial NOT NULL,
     login VARCHAR (200) NOT NULL,
-    pw VARCHAR (64) NOT NULL,
+    pw VARCHAR (128) NOT NULL,
     title VARCHAR (50) NULL,
     first_name VARCHAR (100) NOT NULL,
     last_name VARCHAR (100) NOT NULL,
@@ -1674,7 +1674,7 @@ CREATE TABLE customer_user (
     login VARCHAR (200) NOT NULL,
     email VARCHAR (150) NOT NULL,
     customer_id VARCHAR (150) NOT NULL,
-    pw VARCHAR (64) NULL,
+    pw VARCHAR (128) NULL,
     title VARCHAR (50) NULL,
     first_name VARCHAR (100) NOT NULL,
     last_name VARCHAR (100) NOT NULL,