+++ scripts/test/CustomerAuth.t
@@ -103,7 +103,7 @@ my @Tests = (
     },
 );
 
-for my $CryptType (qw(plain crypt apr1 md5 sha1 sha2 bcrypt)) {
+for my $CryptType (qw(plain crypt apr1 md5 sha1 sha2 sha512 bcrypt)) {
 
     # make sure that the customer user objects gets recreated for each loop.
     $Kernel::OM->ObjectsDiscard(